analog mapping return float value now noize parameter working against input of mapping translation

+++ lighthub/inputs.h
@@ -103,7 +103,7 @@ public:
     boolean isValid();
 
     void onContactChanged(int newValue);
-    void onAnalogChanged(int newValue);
+    void onAnalogChanged(float newValue);
 
     int poll(short cause);
     void setup();
